Restructure the files and redesign some parts of the GUI interface
